Make :editor and :game two explicit modes, remove URL changes
This frees up pages to be an editor-only concept.

With tabs in place we no longer expose the current URL publicly. We
still use it to create identifiable keys for open pages but public URLs
were never that useful since they referenced entities from a given save
file and without that particular save file, the URLs point nowhere.