Do not separate lines only to merge them later
Instead of making the drag-canvas component take a list of component/ids
pairs, we can simple switch on the lines kind to determine the line
component.

Then we don't need to split them in the subscription in the first place
but can simply pass the line ids with their respective kind.